requirements.txt
wheel
tqdm
lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy==1.0.3
imageio
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
scikit-video
opencv-python-headless
roma
git+https://github.com/facebookresearch/pytorch3d.git